moved function render point
